requirements.txt
tensorflow-cpu
flask
gunicorn
colour-science
numpy
base64
skimage
PIL